chore: refactor non-inlinable class extraction to a single function
